+++ src/index.js
@@ -1,8 +1,0 @@
-// @flow
-
-import Replacer from './Replacer';
-
-export default function battleCasex(text: string, name: ?string) {
-  if (!name) return text;
-  return new Replacer(name, text).replaceAll().text;
-}