+++ src/types/casex-template-pattern.ts
@@ -1,0 +1,6 @@
+export type CasexTemplatePatternType = "static" | "singular" | "plural";
+
+export interface CasexTemplatePattern {
+	readonly type: CasexTemplatePatternType;
+	readonly casexPattern: string;
+}