+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 4.0.0
+
+* Casex Template is now part of the DXTR.DEV umbrella
+* Updated function signature to use named arguments for better clarity
+* Use ESM instead of UMD
+* Update from default to named export
+* Converted from Flow to Typescript
+
 # 1.0.0
 
 - Export only battleCasex for simplicity